Vue env and icons

+++ mailserver-configurator-client/src/views/TLSPolicy.vue
@@ -1,7 +1,7 @@
 <template>
     <div>
         <v-container>
-            <v-card>
+            <v-card style="padding-bottom: 10px;">
                 <v-card-title>
                     TLSPolicy
                     <v-spacer></v-spacer>
@@ -15,6 +15,11 @@
 
                     ></v-text-field>
                 </v-card-title>
+                <span style="background-color:#BBDEFB; margin-left: 10px; border-radius: 5px; padding-top: 10px;padding-bottom:8px;">
+                    <v-btn to="/tls/new" icon><v-icon>mdi-plus-circle-outline</v-icon></v-btn>
+                    <v-btn @click="removePolicy()" v-if="selected[0]" icon><v-icon>mdi-close-circle-outline</v-icon></v-btn>
+                    <v-btn @click="editPolicy()" v-if="selected[0]" icon><v-icon>mdi-circle-edit-outline</v-icon></v-btn>
+                </span>
                 <v-data-table
                         :headers="headers"
                         :items="tlspolicys"
@@ -23,9 +28,12 @@
                         v-model="selected"
                         show-select
                 ></v-data-table>
-                <v-btn @click="removePolicy()" v-if="selected[0]">Remove selected Policy</v-btn>
-                <v-btn @click="editPolicy()" v-if="selected[0]">Edit TLS Policy</v-btn>
-                <v-btn to="/tls/new">New TLS-Policy</v-btn><br><br>
+                <span style="background-color:#BBDEFB; margin-left: 10px; border-radius: 5px; padding-top: 10px;padding-bottom:8px;">
+                    <v-btn to="/tls/new" icon><v-icon>mdi-plus-circle-outline</v-icon></v-btn>
+                    <v-btn @click="removePolicy()" v-if="selected[0]" icon><v-icon>mdi-close-circle-outline</v-icon></v-btn>
+                    <v-btn @click="editPolicy()" v-if="selected[0]" icon><v-icon>mdi-circle-edit-outline</v-icon></v-btn>
+                </span>
+
             </v-card>
 
 
@@ -53,9 +61,18 @@
                 this.$router.push("/tls/"+this.selected[0].id)
             },
             removePolicy: function () {
-                Client.deleteTLSPolicy(this.selected[0].id).then(() => {
-                   this.getPolicys();
+                this.$swal({
+                    'title': 'Delete TLS Policy',
+                    'icon': "warning",
+                    showCancelButton: true,
+                }).then((res) => {
+                    if(res.value) {
+                        Client.deleteTLSPolicy(this.selected[0].id).then(() => {
+                            this.getPolicys();
+                        });
+                    }
                 });
+
             }
         },
         mounted: function() {
